Refactored to move reviews and ratings state up to parent component
