+++ client/src/components/Ratings&Reviews/RatingBreakdown.jsx
@@ -1,28 +1,13 @@
 import React, {useState, useEffect, useContext} from "react"
-import axios from "axios"
-import AppContext from '../App/AppContext.jsx'
 import RatingsAndReviewsContext from './RatingsandReviewsContext.jsx'
 import AvgStarRating from "./AvgStarRating.jsx"
 import BreakDown from "./BreakDown.jsx"
 
 const RatingBreakdown = () => {
-  const { product } = useContext(AppContext)
-  const { product_id } = useContext(RatingsAndReviewsContext)
-  const [metaData, setMetaData] = useState({})
+  const { metaData } = useContext(RatingsAndReviewsContext)
   const averageRating = getAvgRating(metaData.ratings)
   const percentRecommend = getPerRecommend(metaData.recommended)
   const breakDown = getBreakDown(metaData.ratings)
-
-  function fetchMetaData() {
-    axios.get(`/api/reviews/meta/?product_id=${product_id}`)
-      .then(response => setMetaData(response.data))
-  }
-
-  useEffect(() => {
-    if(JSON.stringify(product) !== '{}') {
-      fetchMetaData()
-    }
-  }, [product])
 
   function getAvgRating(ratings) {
     let totalStars = 0
